refactor create update to custom hook
