+++ src/module/hook/update/index.ts
@@ -1,4 +1,4 @@
-import { useQuery } from 'react-query';
+import { useQuery, useMutation, useQueryClient } from 'react-query';
 import { IUpdateQuery } from 'module/interface/update';
 import * as api from 'module/api';
 
@@ -14,5 +14,21 @@ export function useUpdates(updateQuery: IUpdateQuery) {
     ['update', updateQuery.timelineId],
     () => getUpdates(updateQuery),
     { enabled: !!updateQuery.timelineId },
+  );
+}
+
+export function useCreateUpdate() {
+  const queryClient = useQueryClient();
+
+  return useMutation<any, any, any>(
+    (newUpdate) => {
+      return api.API.TexUpd.Create(newUpdate);
+    },
+    {
+      onSuccess: () => {
+        // Invalidate and refetch
+        queryClient.invalidateQueries('update');
+      },
+    },
   );
 }
